requirements.txt
streamlit
transformers
pymupdf
pillow
scikit-learn
pytesseract
pandas
numpy
matplotlib
seaborn
torch
scikit-learn
joblib
pyvisa
pymodbus
plotly
opencv-python
PyPDF2
